refactor(app): 실시간 급상승코드를 RealtimeGiftRank 컴포넌트로 분리함
